+++ apps/mobile/src/components/common/search.tsx
@@ -1,17 +1,25 @@
+import {
+  type StyleProp,
+  type TextInputProps,
+  type ViewStyle,
+} from 'react-native'
 import { createStyleSheet, useStyles } from 'react-native-unistyles'
 import { useTranslations } from 'use-intl'
 
 import { Icon } from '~/components/common/icon'
-import { IconButton } from '~/components/common/icon-button'
 import { TextBox } from '~/components/common/text-box'
 
+import { IconButton } from './icon-button'
+
 type Props = {
-  onChange: (value: string) => void
-  value: string
+  onChange?: (value: string) => void
+  onSubmitEditing?: TextInputProps['onSubmitEditing']
+  style?: StyleProp<ViewStyle>
+  value?: string
 }
 
-export function CommunitySearchBox({ onChange, value }: Props) {
-  const t = useTranslations('component.communities.searchBox')
+export function SearchBox({ onChange, onSubmitEditing, style, value }: Props) {
+  const t = useTranslations('component.common.search')
 
   const { styles, theme } = useStyles(stylesheet)
 
@@ -28,10 +36,11 @@ export function CommunitySearchBox({ onChange, value }: Props) {
         />
       }
       onChangeText={onChange}
+      onSubmitEditing={onSubmitEditing}
       placeholder={t('placeholder')}
       returnKeyType="search"
       right={
-        value.length > 0 ? (
+        value?.length ? (
           <IconButton
             icon={{
               color: 'gray',
@@ -39,19 +48,24 @@ export function CommunitySearchBox({ onChange, value }: Props) {
               weight: 'fill',
             }}
             onPress={() => {
-              onChange('')
+              onChange?.('')
             }}
+            style={styles.clear}
           />
         ) : null
       }
-      style={styles.main}
+      style={[styles.main, style]}
       styleContent={styles.content}
       value={value}
     />
   )
 }
 
-const stylesheet = createStyleSheet((theme, runtime) => ({
+const stylesheet = createStyleSheet((theme) => ({
+  clear: {
+    height: theme.space[7],
+    width: theme.space[7],
+  },
   content: {
     backgroundColor: 'transparent',
     borderWidth: 0,
@@ -60,7 +74,7 @@ const stylesheet = createStyleSheet((theme, runtime) => ({
     marginLeft: theme.space[3],
   },
   main: {
+    flexGrow: 1,
     height: theme.space[8],
-    marginTop: runtime.insets.top,
   },
 }))